Move PersistentVolume logic to its own module
